@@ -3,24 +3,23 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.24.0",
-    "carbon-components": "^10.50.0",
-    "carbon-components-react": "^7.50.0",
-    "carbon-icons": "^7.0.7",
-    "http-proxy-middleware": "^2.0.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "axios": "^1.7.9",
+    "@carbon/react": "^1.68.2",
+    "@carbon/icons-react": "^11.54.0",
+    "http-proxy-middleware": "^2.0.6",
     "js-file-download": "^0.4.12",
     "js-yaml": "^4.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "sass": "^1.47.0",
-    "web-vitals": "^2.1.3"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.83.4",
+    "web-vitals": "^4.2.4"
   },
   "overrides": {
-    "@svgr/webpack": "^6.2.1"
+    "@svgr/webpack": "^8.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
