@@ -2,7 +2,7 @@ const { createProxyMiddleware } = require('http-proxy-middleware');
 
 module.exports = function(app) {
   app.use(
-    '/api/v1/*',
+    '/api/v1',
     createProxyMiddleware({
       target: 'http://localhost:32080',
       changeOrigin: true,
